LayoutTests/platform/chromium/inspector/console-log-native-funcs-expected.txt

CONSOLE MESSAGE: line 9: function Document() { [native code] }
CONSOLE MESSAGE: line 10: function createElement() { [native code] }
Tests that console logging dumps proper messages.

console-log-native-funcs.html:9function Document() { [native code] } console-message console-js-source console-log-level
console-log-native-funcs.html:10function createElement() { [native code] } console-message console-js-source console-log-level
